chore: move datagrid skeleton story to base component
